+++ mobile-flutter/lib/features/pengaturan/data/repositories/pengaturan_repository_impl.dart
@@ -1,5 +1,6 @@
 import '../../../../core/network/api_exception.dart';
 import '../../../../core/network/api_result.dart';
+import '../../domain/entities/pengaturan_pembayaran.dart';
 import '../../domain/entities/profil_usaha.dart';
 import '../../domain/repositories/pengaturan_repository.dart';
 import '../datasources/pengaturan_remote_datasource.dart';
@@ -13,6 +14,15 @@ class PengaturanRepositoryImpl implements PengaturanRepository {
   Future<Result<ProfilUsaha>> profilUsaha() async {
     try {
       return Ok(await remote.profilUsaha());
+    } on ApiException catch (e) {
+      return Err(e);
+    }
+  }
+
+  @override
+  Future<Result<PengaturanPembayaran>> pembayaran() async {
+    try {
+      return Ok(await remote.pembayaran());
     } on ApiException catch (e) {
       return Err(e);
     }